Implement #38: Extract extractKey/extractRepo to internal/urlutil package
